rename `settingTab.ts` file to `StickyNotesSettingsTab.ts`
