+++ core/views/StickyNotesSettingsTab.ts
